C++ Parallelization of The Discrete Fourier Transformation using MPI
This is a C++ program that Parallelizes the DFT using MPI(Message Passing Interface). This was One of my assignments for my High Performance Computing class. This program was then used to gather timing data for a set amount of inputs with different numbers of 'workers' to identify the optimal performance point.